API Reference

RayTracingBasics.DirectionVector — Type

DirectionVector{FT<:AbstractFloat} <: FieldVector{3,FT}
DirectionVector(x::FT, y::FT, z::FT) where {FT<:AbstractFloat}

Unit-length vector representing direction in 3D space.

Constructor arguments

x, y, z: direction components [-]

Fields

xd, yd, zd: normalised direction components [-]

RayTracingBasics.DirectionVector — Method

DirectionVector(alpha::FT, beta::FT) where {FT<:AbstractFloat}

Constructor for DirectionVector using spherical coordinates.

Arguments

alpha: elevation angle [radians]

beta: azimuth angle [radians]

Returns

DirectionVector{FT} constructed from spherical coordinates

RayTracingBasics.PositionVector — Type

PositionVector{FT<:AbstractFloat} <: FieldVector{3,FT}
PositionVector(xc::FT, yc::FT, zc::FT) where {FT<:AbstractFloat}

Vector representing position in 3D space.

Constructor arguments

xc, yc, zc: position components [metres]

Fields

Same as constructor arguments

RayTracingBasics.t_at_plane — Method

t_at_plane(ray_origin::PositionVector{FT}, ray_direction::DirectionVector{FT}, axis::Int, c::FT) where {FT<:AbstractFloat}

Compute the ray parameter t at which a ray intersects an axis-aligned plane.

Arguments

ray_origin, ray_direction: ray origin [metres] and direction [-]

axis: axis normal to the plane (1=x, 2=y, 3=z)

c: coordinate of the plane along axis [metres]

Returns

FT: ray parameter t at the intersection point [metres]

RayTracingBasics.to_homogeneous — Method

to_homogeneous(v::DirectionVector{FT}) where {FT<:AbstractFloat}

Convert a DirectionVector to homogeneous coordinates by appending a zero weight.

Arguments

v: direction vector in 3D space

Returns

SVector{4, FT}: homogeneous coordinate vector [xd, yd, zd, 0]

RayTracingBasics.to_homogeneous — Method

to_homogeneous(v::PositionVector{FT}) where {FT<:AbstractFloat}

Convert a PositionVector to homogeneous coordinates by appending a unit weight.

Arguments

v: position vector in 3D space

Returns

SVector{4, FT}: homogeneous coordinate vector [xc, yc, zc, 1]

StaticArraysCore.similar_type — Method

StaticArrays.similar_type(::Type{<:DirectionVector}, ::Type{T}, s::Size{S}) where {T,S}

Extend StaticArrays.similar_type to return a standard SVector when doing maths on a DirectionVector. This is because the result of arithmetic on a DirectionVector is not generally a unit vector.